backend/requirements.txt
Flask
PyPDF2
Werkzeug
reportlab
flask-cors
Pillow
gunicorn
flask-jwt-extended
flask-bcrypt
psycopg2
python-dotenv